Create game teams mangager class
